Don't show identical adapters
The code tries to coalesce identical adapters but when
chrome's webgpu-developer-features is on, then powerPreference
shows up in an adapter's id and it shows two adapters.

Maybe that is ok as it only affects people with that flag on?

My original goal here was to print which is the default adapter.
I've added that by requesting one adapter with no preferences.
It should match either 'low-power' or 'high-performance'. It's
always the first adapter so we can check if its id matches
to find the default.

When powerPreference got added to adapterInfo